conversation/requirements.txt
python-dotenv
semantic-kernel
chainlit
fastmcp
